phishing-detector/requirements.txt
requests==2.32.3
tldextract==5.1.2
python-Levenshtein==0.26.0
Pillow==10.4.0